move integration tests to separate directory, configure paths in tsconfig
