+++ .gitignore
@@ -1,4 +1,4 @@
 .nyc_output
 coverage
-lib/__tests__/contracts/*.bin
-lib/__tests__/contracts/*.abi
+node_modules
+test/contracts-compiled
