+++ lib/opcodes/__tests__/arithmetic.spec.ts
@@ -1,4 +1,4 @@
-import { compareWithReferentialImpl } from "../../__tests__/helpers/compareWithReferentialImpl";
+import { compareWithReferentialImpl } from "test/helpers/compareWithReferentialImpl";
 
 describe("arithmetic opcodes", () => {
   describe("ADD", () => {
