+++ lib/opcodes/__tests__/memory.spec.ts
@@ -1,4 +1,4 @@
-import { compareWithReferentialImpl } from "../../__tests__/helpers/compareWithReferentialImpl";
+import { compareWithReferentialImpl } from "test/helpers/compareWithReferentialImpl";
 
 describe("memory opcodes", () => {
   describe("MSTORE", () => {
