+++ lib/opcodes/__tests__/push.spec.ts
@@ -1,4 +1,4 @@
-import { compareWithReferentialImpl } from "../../__tests__/helpers/compareWithReferentialImpl";
+import { compareWithReferentialImpl } from "test/helpers/compareWithReferentialImpl";
 
 describe("PUSH", () => {
   it("should push simple data", () =>
