+++ tsconfig.json
@@ -4,7 +4,11 @@
     "strict": true,
     "module": "commonjs",
     "target": "es2015",
-    "allowJs": true
+    "allowJs": true,
+    "baseUrl": ".",
+    "paths": {
+      "test": ["test/*"]
+    }
   },
   "include": ["**/*.ts"]
 }
